LayoutTests/js/dom/global-constructors-attributes-dedicated-worker-expected.txt

Test to ensure that global constructors have the right attributes in the dedicated worker context

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: script-tests/global-constructors-attributes.js
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').value is Blob
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'Blob').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').value is DedicatedWorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'DedicatedWorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').value is EventSource
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'EventSource').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').value is FileReader
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReader').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').value is FileReaderSync
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'FileReaderSync').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursor').value is IDBCursor
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursor').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursor').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursor').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursor').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').value is IDBCursorWithValue
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBDatabase').value is IDBDatabase
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBDatabase').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBDatabase').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBDatabase').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBDatabase').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBFactory').value is IDBFactory
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBFactory').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBFactory').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBFactory').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBFactory').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBIndex').value is IDBIndex
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBIndex').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBIndex').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBIndex').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBIndex').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').value is IDBKeyRange
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').value is IDBObjectStore
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').value is IDBOpenDBRequest
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBRequest').value is IDBRequest
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBRequest').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBRequest').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBRequest').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBRequest').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBTransaction').value is IDBTransaction
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBTransaction').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBTransaction').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBTransaction').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBTransaction').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').value is IDBVersionChangeEvent
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').value is ImageData
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'ImageData').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').value is MessageChannel
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageChannel').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').value is MessageEvent
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'MessageEvent').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').value is URL
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'URL').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').value is WebSocket
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WebSocket').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').value is WorkerGlobalScope
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerGlobalScope').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').value is WorkerLocation
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'WorkerLocation').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').value is XMLHttpRequest
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'XMLHttpRequest').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBCursor').value is webkitIDBCursor
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBCursor').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBCursor').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBCursor').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBCursor').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBDatabase').value is webkitIDBDatabase
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBDatabase').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBDatabase').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBDatabase').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBDatabase').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBFactory').value is webkitIDBFactory
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBFactory').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBFactory').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBFactory').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBFactory').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBIndex').value is webkitIDBIndex
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBIndex').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBIndex').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBIndex').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBIndex').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBKeyRange').value is webkitIDBKeyRange
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBKeyRange').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBKeyRange').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBKeyRange').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBKeyRange').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBObjectStore').value is webkitIDBObjectStore
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBObjectStore').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBObjectStore').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBObjectStore').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBObjectStore').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBRequest').value is webkitIDBRequest
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBRequest').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBRequest').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBRequest').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBRequest').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBTransaction').value is webkitIDBTransaction
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBTransaction').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBTransaction').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBTransaction').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitIDBTransaction').configurable is true
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').value is webkitURL
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').hasOwnProperty('get') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').hasOwnProperty('set') is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').enumerable is false
PASS [Worker] Object.getOwnPropertyDescriptor(global, 'webkitURL').configurable is true
PASS successfullyParsed is true

TEST COMPLETE
